pyproject.toml
[project]
name = "azure-apic-a2a-registry-web-api-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "azure-identity>=1.25.1,<2.0",
    "azure-mgmt-apicenter>=1.0.0,<2.0",
    "flask[async]>=3.1.2,<4.0",
    "lagom>=2.7.7,<3.0",
    "pydantic-settings>=2.12.0,<3.0",
    "pydantic>=2.12.5,<3.0",
    "python-dotenv>=1.2.1,<2.0",
    "uritemplate>=4.2.0,<5.0",
]

[dependency-groups]
dev = [
    "pyright>=1.1.407,<2.0",
    "ruff>=0.14.9,<0.15.0",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "src/a2a_registry",
]
